+++ src/Core/Actions.h
@@ -18,7 +18,7 @@ namespace ESPExplorerAE
         Give, GiveWithAmmo, Spawn, Place, AddSpell, RemoveSpell, AddPerk,
         RemovePerk, Outfit, ConstructedItem, CurrentAmmo, GodMode, Console, Teleport, Equip,
         StartQuest, CompleteQuest, SetWeather, PlaySound, SetGlobal,
-        InventoryRemove, InventoryDrop, InventoryEquip, InventoryUnequip, InventoryUse, InventoryAddBase,
+        InventoryRemove, InventoryDrop, InventoryEquip, InventoryUnequip, InventoryUse, InventoryAddBase, InventoryDuplicateWeapon,
         RestoreHealth, ToggleNoClip, SetPlayerLevel, AddPerkPoints, SetGameHour
     };
 
@@ -50,6 +50,7 @@ namespace ESPExplorerAE
         case ActionKind::InventoryEquip: return InventoryAction::Equip;
         case ActionKind::InventoryUnequip: return InventoryAction::Unequip;
         case ActionKind::InventoryUse: return InventoryAction::Use;
+        case ActionKind::InventoryDuplicateWeapon: return InventoryAction::DuplicateWeapon;
         default: return {};
         }
     }
@@ -62,6 +63,7 @@ namespace ESPExplorerAE
         case InventoryAction::Equip: return ActionKind::InventoryEquip;
         case InventoryAction::Unequip: return ActionKind::InventoryUnequip;
         case InventoryAction::Use: return ActionKind::InventoryUse;
+        case InventoryAction::DuplicateWeapon: return ActionKind::InventoryDuplicateWeapon;
         default: return static_cast<ActionKind>(-1);
         }
     }